Stable-Fast-3D/uv_unwrapper/uv_unwrapper/__init__.py
import torch  # noqa: F401

try:
    from . import _C  # noqa: F401
except ImportError:
    # Extension missing or failed to load; Unwrapper still imports but torch.ops will fail at runtime.
    pass

from .unwrap import Unwrapper

__all__ = ["Unwrapper"]
